SPACE-55: First pass at updating CSS to tailwindcss classes
Updated all the CSS to use tailwindcss classes. Also refactored some of the component structure and moved some files around.
