--- frontend/src/component/ChoosePlanet.css
@@ -1,61 +1,0 @@
-/* Background overlay */
-.popup-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.6); /* Semi-transparent black background */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-/* Popup styles */
-.popup {
-  background-color: #7300ff; /* A soft pink background */
-  padding: 20px;
-  border-radius: 12px; /* Rounded corners */
-  width: 60%;
-  max-width: 500px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
-  text-align: center;
-  color: white;
-  overflow-y: auto;
-  height: 70%;
-}
-
-/* Title */
-.popup-title {
-  font-size: 2rem;
-  font-weight: bold;
-  margin-bottom: 20px;
-}
-
-/* Planet List */
-.planet-list {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-/* Individual Planet Item */
-.planet-item {
-  padding: 10px;
-  background-color: rgba(0, 0, 0, 0.3); /* Dark background for each item */
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.planet-item:hover {
-  background-color: rgba(0, 0, 0, 0.5); /* Slight darkening on hover */
-}
-
-/* Responsive design */
-@media (max-width: 600px) {
-  .popup {
-    width: 80%;
-  }
-}
