+++ frontend/src/component/FuelStatus.jsx
@@ -5,7 +5,7 @@ import TradeContext from "../context/tradeContext";
 function FuelStatus() {
   const { fuel, points } = useContext(TradeContext);
   return (
-    <div style={statusStyle}>
+    <div className="fixed top-5 left-5 bg-black text-white px-6 py-6 rounded-lg z-[1000]">
       <p>
         🚀 Current Fuel: <strong>{fuel}</strong>
       </p>
@@ -15,17 +15,5 @@ function FuelStatus() {
     </div>
   );
 }
-
-const statusStyle = {
-  position: "fixed",
-  top: "10px",
-  left: "10px",
-  backgroundColor: "black",
-  color: "white",
-  padding: "10px 15px",
-  borderRadius: "10px",
-  zIndex: 1000,
-  fontFamily: "sans-serif",
-};
 
 export default FuelStatus;
