+++ frontend/src/component/PlanetJourney.jsx
@@ -1,199 +1,128 @@
-import React, { useContext, useState, useRef, useEffect } from "react";
+import React, {useContext, useState, useRef, useEffect} from "react";
 import QuizModal from "./QuizModal";
 import TradeContext from "../context/tradeContext";
-import { getPlanetInfo } from "../planetInfo/getPlanetInfo";
+import {getPlanetInfo} from "../planetInfo/getPlanetInfo";
 import travelSoundFile from "../assets/spaceship-fly.mp3";
-import { planetDetails } from "../planetInfo/planetDetails";
+import {planetDetails} from "../planetInfo/planetDetails";
 
 
-export function PlanetJourney({ selectedPlanet, onExit }) {
-  const [fade, setFade] = useState(true);
-  const [showQuiz, setShowQuiz] = useState(false);
-  const [showSummary, setShowSummary] = useState(true);
-  const [showAbout, setShowAbout] = useState(false);
-  const [volume, setVolume] = useState(0.5);
-  const [selectedFact, setSelectedFact] = useState(null);
+export function PlanetJourney({selectedPlanet, onExit}) {
+    const [fade, setFade] = useState(true);
+    const [showQuiz, setShowQuiz] = useState(false);
+    const [showSummary, setShowSummary] = useState(true);
+    const [showAbout, setShowAbout] = useState(false);
+    const [volume, setVolume] = useState(0.5);
+    const [selectedFact, setSelectedFact] = useState(null);
 
-  const travelSoundRef = useRef(null);
-  const synthRef = useRef(window.speechSynthesis);
+    const travelSoundRef = useRef(null);
+    const synthRef = useRef(window.speechSynthesis);
 
-  useEffect(() => {
-    setTimeout(() => setFade(true), 2000);
-    if (travelSoundRef.current) {
-      travelSoundRef.current.volume = volume;
-      travelSoundRef.current.play().catch((e) => console.log("Autoplay blocked", e));
-    }
-    return () => {
-      travelSoundRef.current?.pause();
+    useEffect(() => {
+        setTimeout(() => setFade(true), 2000);
+        if (travelSoundRef.current) {
+            travelSoundRef.current.volume = volume;
+            travelSoundRef.current.play().catch((e) => console.log("Autoplay blocked", e));
+        }
+        return () => {
+            travelSoundRef.current?.pause();
+        };
+    }, [volume]);
+
+    const {information, funFacts} = getPlanetInfo(selectedPlanet.name);
+
+    const handleFactClick = (fact) => {
+        setSelectedFact(fact);
+        const utterance = new SpeechSynthesisUtterance(fact);
+        synthRef.current.cancel();
+        synthRef.current.speak(utterance);
     };
-  }, [volume]);
 
-  if (!selectedPlanet) {
-    return <div>Loading... 🚀</div>;
-  }
+    const handleAboutClick = () => {
+        setShowAbout(true);
+        setShowSummary(false);
+    };
 
-  const { information, funFacts } = getPlanetInfo(selectedPlanet.name);
+    const handleReturnClick = () => {
+        setShowAbout(false);
+        setShowSummary(true);
+    };
 
-  const handleFactClick = (fact) => {
-    setSelectedFact(fact);
-    const utterance = new SpeechSynthesisUtterance(fact);
-    synthRef.current.cancel();
-    synthRef.current.speak(utterance);
-  };
+    return (
+        <div id="planet-journey" className={`text-center p-5 transition-opacity duration-2000 ${fade ? 'opacity-100' : 'opacity-0'}`}>
+            <audio ref={travelSoundRef} src={travelSoundFile}/>
 
-  const handleAboutClick = () => {
-    setShowAbout(true);
-    setShowSummary(false);
-  };
+            {/*<h2>🚀 Traveling...</h2>*/}
+            <h2 className="text-lg font-bold mb-4">Welcome to {selectedPlanet.name}!</h2>
+            <img
+                src={selectedPlanet.image}
+                alt={selectedPlanet.name}
+                className="w-[450px] h-[250px] rounded-[10%] mx-auto"
+            />
 
-  const handleReturnClick = () => {
-    setShowAbout(false);
-    setShowSummary(true);
-  };
+            {showSummary && (
+                <div>
+                    <h3 className="mt-6 text-xl font-semibold">🌍 Planet Summary</h3>
+                    <p>{information}</p>
+                    <h3 className="mt-4 text-xl font-semibold">🪐 Fun Planet Facts</h3>
+                    <div className="grid grid-cols-2 gap-4 justify-items-center mt-4">
+                        {funFacts?.map((fact, index) => {
+                            const isSelected = selectedFact === fact;
+                            const isThreeFactsMiddle = funFacts.length === 3 && index === 2;
+                            return (
+                                <div
+                                    key={index}
+                                    onClick={() => handleFactClick(fact)}
+                                    className={`${
+                                        isThreeFactsMiddle ? 'col-span-2' : ''
+                                    } max-w-[300px] px-6 py-4 rounded-2xl text-center text-base font-medium leading-snug cursor-pointer transform transition-all duration-300 shadow-md ${
+                                        isSelected
+                                            ? 'bg-gradient-to-br from-blue-600 to-blue-900 text-white scale-105'
+                                            : 'bg-gradient-to-br from-gray-800 to-gray-900 text-gray-200 hover:scale-105'
+                                    }`}
+                                >
+                                    {fact}
+                                </div>
+                            );
+                        })}
+                    </div>
+                    <div className="mt-6 flex justify-center gap-4">
+                        <button
+                            onClick={() => setShowQuiz(true)}
+                            className="p-2 px-6 bg-blue-600 hover:bg-blue-800 disabled:bg-zinc-600 text-base text-white disabled:text-zinc-400 mb-2 rounded-lg transition">
+                            Start Quiz
+                        </button>
+                        <button
+                            onClick={handleAboutClick}
+                            className="p-2 px-6 bg-blue-600 hover:bg-blue-800 disabled:bg-zinc-600 text-base text-white disabled:text-zinc-400 mb-2 rounded-lg transition">
+                            Explore More
+                        </button>
+                    </div>
+                </div>
+            )}
 
-  return (
-    <div
-      style={{
-        textAlign: "center",
-        padding: "20px",
-        opacity: fade ? 1 : 0,
-        transition: "opacity 2s",
-      }}
-    >
-      <audio ref={travelSoundRef} src={travelSoundFile} />
+            {showAbout && (
+                <div className="mt-5">
+                    <h3 className="text-xl font-semibold">🧭 About {selectedPlanet.name}</h3>
+                    <p><strong>Terrain:</strong> {planetDetails[selectedPlanet.name]?.terrain}</p>
+                    <p><strong>Atmosphere:</strong> {planetDetails[selectedPlanet.name]?.atmosphere}</p>
+                    <p><strong>Did You Know?</strong> {planetDetails[selectedPlanet.name]?.interestingFact}</p>
+                    <button
+                        onClick={handleReturnClick}
+                        className="mt-5 p-2 px-6 bg-blue-600 hover:bg-blue-800 disabled:bg-zinc-600 text-base text-white disabled:text-zinc-400 mb-2 rounded-lg transition">
+                        Return
+                    </button>
+                </div>
+            )}
 
-      <h2>🚀 Traveling...</h2>
-      <h2>Welcome to {selectedPlanet.name}!</h2>
-      <img
-        src={selectedPlanet.image}
-        alt={selectedPlanet.name}
-        style={{ width: "450px", height: "250px", borderRadius: "10%" }}
-      />
-
-      {showSummary && (
-        <div>
-          <h3>🌍 Planet Summary</h3>
-          <p>{information}</p>
-          <h3>🪐 Fun Planet Facts</h3>
-          <div
-            style={{
-              display: "grid",
-              gridTemplateColumns: "repeat(2, minmax(0, 1fr))",
-              gap: "16px",
-              justifyItems: "center",
-              marginTop: "16px",
-            }}
-          >
-            {funFacts?.map((fact, index) => (
-              <div
-                key={index}
-                onClick={() => handleFactClick(fact)}
-                style={{
-                  gridColumn: funFacts.length === 3 && index === 2 ? "1 / span 2" : "auto",
-                  background: selectedFact === fact
-                    ? "linear-gradient(135deg, #2563eb, #1e3a8a)"
-                    : "linear-gradient(135deg, #1f2937, #111827)",
-                  color: selectedFact === fact ? "#ffffff" : "#e5e7eb",
-                  borderRadius: "16px",
-                  padding: "16px 24px",
-                  maxWidth: "300px",
-                  cursor: "pointer",
-                  boxShadow: selectedFact === fact
-                    ? "0 0 12px rgba(37, 99, 235, 0.7)"
-                    : "0 2px 8px rgba(0, 0, 0, 0.25)",
-                  transition: "all 0.3s ease",
-                  transform: selectedFact === fact ? "scale(1.08)" : "scale(1)",
-                  fontSize: "16px",
-                  fontWeight: 500,
-                  lineHeight: "1.4",
-                  textAlign: "center",
-                }}
-                onMouseEnter={(e) => (e.currentTarget.style.transform = "scale(1.05)")}
-                onMouseLeave={(e) =>
-                  (e.currentTarget.style.transform =
-                    selectedFact === fact ? "scale(1.08)" : "scale(1)")
-                }
-              >
-                {fact}
-              </div>
-            ))}
-          </div>
-          <div style={{ marginTop: "24px", display: "flex", justifyContent: "center", gap: "16px" }}>
-            <button
-              onClick={() => setShowQuiz(true)}
-              style={{
-                padding: "10px 20px",
-                fontSize: "16px",
-                borderRadius: "8px",
-                background: "#1e40af",
-                color: "#fff",
-                border: "none",
-                cursor: "pointer",
-                transition: "background 0.3s",
-              }}
-              onMouseEnter={(e) => (e.currentTarget.style.background = "#1e3a8a")}
-              onMouseLeave={(e) => (e.currentTarget.style.background = "#1e40af")}
-            >
-              Start Quiz
-            </button>
-            <button
-              onClick={handleAboutClick}
-              style={{
-                padding: "10px 20px",
-                fontSize: "16px",
-                borderRadius: "8px",
-                background: "#1e40af",
-                color: "#fff",
-                border: "none",
-                cursor: "pointer",
-                transition: "background 0.3s",
-              }}
-              onMouseEnter={(e) => (e.currentTarget.style.background = "#1e3a8a")}
-              onMouseLeave={(e) => (e.currentTarget.style.background = "#1e40af")}
-            >
-              Explore More
-            </button>
-          </div>
+            {showQuiz && selectedPlanet && (
+                <QuizModal
+                    selectedPlanet={selectedPlanet}
+                    onClose={() => {
+                        setShowQuiz(false);
+                        onExit();
+                    }}
+                />
+            )}
         </div>
-      )}
-
-{showAbout && (
-  <div style={{ marginTop: "20px" }}>
-    <h3>🧭 About {selectedPlanet.name}</h3>
-    <p><strong>Terrain:</strong> {planetDetails[selectedPlanet.name]?.terrain}</p>
-    <p><strong>Atmosphere:</strong> {planetDetails[selectedPlanet.name]?.atmosphere}</p>
-    <p><strong>Did You Know?</strong> {planetDetails[selectedPlanet.name]?.interestingFact}</p>
-    <button
-      onClick={handleReturnClick}
-      style={{
-        marginTop: "20px",
-        padding: "10px 20px",
-        fontSize: "16px",
-        borderRadius: "8px",
-        background: "#374151",
-        color: "#fff",
-        border: "none",
-        cursor: "pointer",
-        transition: "background 0.3s",
-      }}
-      onMouseEnter={(e) => (e.currentTarget.style.background = "#1f2937")}
-      onMouseLeave={(e) => (e.currentTarget.style.background = "#374151")}
-    >
-      Return
-    </button>
-  </div>
-)}
-
-
-      {showQuiz && selectedPlanet && (
-        <QuizModal
-          selectedPlanet={selectedPlanet}
-          onClose={() => {
-            setShowQuiz(false);
-            onExit();
-          }}
-        />
-      )}
-    </div>
-  );
+    );
 }
