--- frontend/src/component/SpaceStation.css
@@ -1,137 +1,0 @@
-.space-station {
-    position: fixed;
-    top: 0px;
-    bottom: 0px;
-    left: 0px;
-    right: 0px;
-    background: linear-gradient(to bottom, rgba(0, 0, 0, 0.75) 0%, rgba(10, 10, 10, 0.75) 100%);
-    color: #00ffcc;
-    font-family: 'Orbitron', sans-serif;
-    z-index: 9999;
-    padding: 40px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-start;
-    gap: 30px;
-    border: 2px solid rgba(0, 255, 255, 0.4);
-    box-shadow: 0 0 30px rgba(0, 255, 255, 0.2);
-    overflow: hidden;
-    animation: flicker 1s infinite;
-}
-
-/* Panel to hold trade controls */
-.trade-panel {
-    background: rgba(20, 20, 20, 0.95);
-    border: 2px solid #00ffcc;
-    border-image: linear-gradient(to right, #00ffcc, #006666) 1;
-    box-shadow: inset 0 0 10px rgba(0, 255, 255, 0.2);
-    padding: 24px;
-    /*width: 100%;*/
-    /*max-width: 500px;*/
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    /*clip-path: polygon(0 0, 100% 0, 100% 92%, 90% 100%, 0 100%);*/
-}
-
-.resource-list {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    margin: 0px;
-    padding: 0px;
-}
-
-/* Resource listing rows */
-.resource-list .resource {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: rgba(255, 255, 255, 0.07);
-    padding: 10px 14px;
-    border-radius: 6px;
-    font-family: 'Share Tech Mono', monospace;
-    font-size: 1rem;
-    color: #0ff;
-}
-
-.resource button {
-    background-color: #0ff;
-    border: none;
-    padding: 6px 12px;
-    color: #000;
-    font-weight: bold;
-    font-family: inherit;
-    cursor: pointer;
-    border-radius: 4px;
-    transition: background-color 0.2s ease;
-}
-
-.resource button:hover:not(:disabled) {
-    background-color: #00e6cc;
-}
-
-.resource button:disabled {
-    background-color: #333;
-    color: #777;
-    cursor: not-allowed;
-}
-
-/* Fuel display area */
-.fuel-display {
-    font-size: 1.2rem;
-    text-align: center;
-    font-family: 'Share Tech Mono', monospace;
-    text-shadow: 0 0 6px rgba(0,255,255,0.3);
-}
-
-.fuel-value {
-    padding: 0px 20px;
-}
-
-/* Exit button */
-.exit-button {
-    margin-top: auto;
-    background-color: transparent;
-    border: 2px solid #0ff;
-    color: #0ff;
-    padding: 12px 24px;
-    font-size: 1rem;
-    border-radius: 8px;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-    font-family: 'Orbitron', sans-serif;
-}
-
-.exit-button:hover {
-    background-color: #00FFFF;
-    border: 2px solid #004444;
-    color: #004444;
-}
-
-/* Enter button */
-.enter-button {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    margin-top: auto;
-    background-color: black;
-    border: 2px solid #0ff;
-    color: #0ff;
-    padding: 12px 24px;
-    font-size: 1rem;
-    border-radius: 8px;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-    font-family: 'Orbitron', sans-serif;
-}
-
-.enter-button:hover {
-    background-color: #00FFFF;
-    border: 2px solid #004444;
-    color: #004444;
-}
-
-.top-left { top: 20px; left: 20px; }
-.bottom-right { bottom: 20px; right: 20px; }
