+++ frontend/src/component/TradeResource.jsx
@@ -27,17 +27,14 @@ export function TradeResource({resource}) {
     }
 
     return (
-        <li className="resource" style={{opacity: resource.available ? 1 : 0.5}}>
-            <div className="resource-name" data-testid="trade-resource-name">
-                {resource?.name}
-            </div>
-            <div>
-                <span className="fuel-value">+{resource?.value} Fuel</span>
-                <button
-                    onClick={onTrade}
-                    disabled={!resource.available}
-                    data-testid="trade-terminal-trade-button"
-                >
+        <li className="flex justify-between items-center bg-white/10 text-green-300 text-base font-mono px-4 py-2 rounded-md">
+            <div>{resource?.name}</div>
+            <div className="flex items-center gap-4">
+                <span className="pr-5">+{resource?.value} Fuel</span>
+                <button onClick={onTrade}
+                        disabled={!resource.available}
+                        data-testid="trade-terminal-trade-button"
+                        className="bg-teal-300 text-black font-bold px-3 py-1 rounded hover:bg-green-400 disabled:bg-neutral-700 disabled:text-gray-400 disabled:cursor-not-allowed">
                     Trade
                 </button>
             </div>
