+++ frontend/src/component/TradeTerminal.jsx
@@ -17,25 +17,23 @@ export function TradeTerminal({selectedPlanet}) {
     }));
 
     return (
-        <div>
-            <h2>Trade Terminal</h2>
-            <p>Exchange resources for fuel. Prices may vary.</p>
+        <div className="flex-auto bg-white/5 border border-green-400 text-green-300 rounded-lg p-4 overflow-y-auto">
+            <div className="text-center space-y-2 mb-5">
+                <h2 className="text-2xl font-bold">Trade Terminal</h2>
+                <p className="text-sm">Exchange resources for fuel. Prices may vary.</p>
+            </div>
 
-            <div className="trade-panel">
-                <ul className="resource-list">
-                    {resources.map((resource, index) => (
-                        <TradeResource
-                            key={index}
-                            resource={resource}
-                        />
-                    ))}
-                </ul>
+            <ul className="space-y-2">
+                {resources.map((resource, index) => (
+                    <TradeResource
+                        key={index}
+                        resource={resource}
+                    />
+                ))}
+            </ul>
 
-                <div className="fuel-display">
-                    <p>
-                        Current Fuel: <strong>{fuel}</strong>
-                    </p>
-                </div>
+            <div className="mt-4 text-center text-lg font-mono text-shadow-sm">
+                Current Fuel: <strong>{fuel}</strong>
             </div>
         </div>
     );
