+++ frontend/src/utils/getElipticLines.js
