+++ frontend/src/utils/getStarField.js
