+++ frontend/src/utils/getSun.js
